Cleanup: Move Vol implementations into Vol.cc
