+++ iocore/cache/Makefile.am
@@ -58,7 +58,8 @@ libinkcache_a_SOURCES = \
 	P_RamCache.h \
 	RamCacheCLFUS.cc \
 	RamCacheLRU.cc \
-	Store.cc
+	Store.cc \
+	Vol.cc
 
 if BUILD_TESTS
 libinkcache_a_SOURCES += \
